// src/app/core/models/user.model.ts
export class User {
    'id': string;
    'client_first_name': string;
    'client_last_name': string;
    'contact_no': string;
    'email': string;
    'profile_pic': string;
    'google_place_id': string;
    'google_place_name': string;
    'device_type': string;
    'fb_login': boolean;
    'is_login': string;
}